REF: move bulk_insert logic to core
