+++ filestore/commands.py
@@ -13,6 +13,8 @@ from jsonschema import validate as js_validate
 
 from document import Document
 
+from .core import bulk_insert_datum as _bulk_insert_datum
+
 
 def _ensure_connection(func):
     @wraps(func)
@@ -113,21 +115,8 @@ def insert_datum(resource, datum_id, datum_kwargs=None):
 
 @_ensure_connection
 def bulk_insert_datum(resource, datum_ids, datum_kwarg_list):
-
-    resource_id = resource['id']
-
-    def datum_factory():
-        for d_id, d_kwargs in zip(datum_ids, datum_kwarg_list):
-            datum = dict(resource=resource_id,
-                         datum_id=d_id,
-                         datum_kwargs=d_kwargs)
-            yield datum
-
-    bulk = Datum._get_collection().initialize_ordered_bulk_op()
-    for dm in datum_factory():
-        bulk.insert(dm)
-
-    return bulk.execute()
+    col = Datum._get_collection()
+    return _bulk_insert_datum(col, resource, datum_ids, datum_kwarg_list)
 
 
 @_ensure_connection
